refactor: 순환 구조 제거 InputViewUtil <--> InputVIew 구조에서 InputViewUtil -> InputView 으로 변경
